fix(Presentor): Changed some formuls

+++ app/src/main/java/com/bei/andrii/forexcalculator/activity/MainActivity.java
@@ -4,9 +4,10 @@ public interface MainActivity {
 
     void showResultForStep(int mPPrice, int mPStop, int mPTake);
 
-    void showResultAmount(long mAmountEnter, float mExpectedProfit, float mExpectedLesion, float mExpectedProfitPercent, float mExpectedLesionPercent);
+    void showResultAmount(float mAmountEnter, double mExpectedProfit, double mExpectedLesion, float mExpectedProfitPercent, float mExpectedLesionPercent);
 
     void showResultComment(int mComment);
 
     void showErrorSnackBarDivideOnZero();
+
 }
